+++ pm/src/main/java/io/reactiverse/es4x/commands/InstallCommand.java
@@ -15,8 +15,6 @@
  */
 package io.reactiverse.es4x.commands;
 
-import com.fasterxml.jackson.core.JsonParser;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.reactiverse.es4x.ES4X;
 import io.vertx.core.cli.CLIException;
 import io.vertx.core.cli.annotations.*;
@@ -37,19 +35,17 @@ import static io.reactiverse.es4x.commands.Helper.*;
 @Summary("Installs required jars from maven to 'node_modules'.")
 public class InstallCommand extends DefaultCommand {
 
-  private static final ObjectMapper MAPPER = new ObjectMapper();
   private static final Properties VERSIONS = new Properties();
 
   static {
-    MAPPER.configure(JsonParser.Feature.ALLOW_COMMENTS, true);
     try (InputStream is = InstallCommand.class.getClassLoader().getResourceAsStream("META-INF/es4x-commands/VERSIONS.properties")) {
       if (is == null) {
-        err("Cannot find 'META-INF/es4x-commands/VERSIONS.properties' on classpath");
+        fatal("Cannot find 'META-INF/es4x-commands/VERSIONS.properties' on classpath");
       } else {
         VERSIONS.load(is);
       }
     } catch (IOException e) {
-      err(e.getMessage());
+      fatal(e.getMessage());
     }
   }
 
@@ -70,7 +66,7 @@ public class InstallCommand extends DefaultCommand {
 
   private static void processPackageJson(File json, Set<String> dependencies) throws IOException {
     if (json.exists()) {
-      Map npm = MAPPER.readValue(json, Map.class);
+      Map npm = read(json);
       if (npm.containsKey("maven")) {
         final Map maven = (Map) npm.get("maven");
         // add this dependency
@@ -134,22 +130,26 @@ public class InstallCommand extends DefaultCommand {
       final double version = Double.parseDouble(System.getProperty("java.specification.version"));
       final String vm = System.getProperty("java.vm.name");
       if (!vm.toLowerCase().contains("graalvm")) {
+
+        // not on graal, install graaljs and dependencies
+        warn("Installing GraalJS...");
+        // graaljs + dependencies
+        installGraalJS(artifacts);
+
         if (version >= 11) {
-          System.err.println("\u001B[1m\u001B[33mInstalling GraalJS on stock JDK!\u001B[0m");
-          // graaljs + dependencies
-          installGraalJS(artifacts);
           // verify if the current JDK contains the jdk.internal.vm.ci module
           try {
             String modules = exec(javaHomePrefix() + "java", "--list-modules");
             if (modules.contains("jdk.internal.vm.ci")) {
+              warn("Installing JVMCI Compiler...");
               // jvmci compiler + dependencies
               installGraalJMVCICompiler();
             }
           } catch (IOException | InterruptedException e) {
-            System.err.println("\u001B[1m\u001B[31m" + e.getMessage() + "\u001B[0m");
+            err(e.getMessage());
           }
         } else {
-          System.err.println("\u001B[1m\u001B[31mCurrent JDK only supports Nashorn!\u001B[0m");
+          warn("Current JDK supports Nashorn [ES5.1] or GraalJS Interpreted!");
         }
       }
     }
@@ -164,7 +164,7 @@ public class InstallCommand extends DefaultCommand {
     File libs = new File(base, ".lib");
     if (!libs.exists()) {
       if (!libs.mkdirs()) {
-        err("Failed to mkdirs 'node_modules/.lib'.");
+        fatal("Failed to mkdirs 'node_modules/.lib'.");
       }
     }
 
@@ -179,7 +179,7 @@ public class InstallCommand extends DefaultCommand {
         }
       }
     } catch (IOException e) {
-      err(e.getMessage());
+      fatal(e.getMessage());
     }
   }
 
@@ -189,7 +189,7 @@ public class InstallCommand extends DefaultCommand {
     File libs = new File(base, ".jvmci");
     if (!libs.exists()) {
       if (!libs.mkdirs()) {
-        err("Failed to mkdirs 'node_modules/.jvmci'.");
+        fatal("Failed to mkdirs 'node_modules/.jvmci'.");
       }
     }
 
@@ -203,7 +203,7 @@ public class InstallCommand extends DefaultCommand {
         }
       }
     } catch (IOException e) {
-      err(e.getMessage());
+      fatal(e.getMessage());
     }
   }
 
@@ -215,7 +215,7 @@ public class InstallCommand extends DefaultCommand {
     try {
       processPackageJson(new File(getCwd(), "package.json"), dependencies);
     } catch (IOException e) {
-      err(e.getMessage());
+      fatal(e.getMessage());
     }
 
     // crawl node modules
@@ -223,7 +223,7 @@ public class InstallCommand extends DefaultCommand {
       try {
         processModules(base, dependencies);
       } catch (IOException e) {
-        err(e.getMessage());
+        fatal(e.getMessage());
       }
     }
 
@@ -248,7 +248,7 @@ public class InstallCommand extends DefaultCommand {
 
           if (!libs.exists()) {
             if (!libs.mkdirs()) {
-              err("Failed to mkdirs 'node_modules/.lib'.");
+              fatal("Failed to mkdirs 'node_modules/.lib'.");
             }
           }
           artifacts.add("../.lib/" + a.getFile().getName());
@@ -258,7 +258,7 @@ public class InstallCommand extends DefaultCommand {
           }
         }
       } catch (IOException e) {
-        err(e.getMessage());
+        fatal(e.getMessage());
       }
     }
   }
@@ -269,13 +269,13 @@ public class InstallCommand extends DefaultCommand {
 
     if (json.exists()) {
       try {
-        Map npm = MAPPER.readValue(json, Map.class);
+        Map npm = read(json);
 
         if (launcher == null) {
           if (npm.containsKey("name")) {
             launcher = (String) npm.get("name");
           } else {
-            err("'package.json' doesn't contain a 'name' property!");
+            fatal("'package.json' doesn't contain a 'name' property!");
           }
         }
 
@@ -291,7 +291,7 @@ public class InstallCommand extends DefaultCommand {
         File bin = new File(base, ".bin");
         if (!bin.exists()) {
           if (!bin.mkdirs()) {
-            err("Failed to mkdirs 'node_modules/.bin'.");
+            fatal("Failed to mkdirs 'node_modules/.bin'.");
           }
         }
 
@@ -320,7 +320,7 @@ public class InstallCommand extends DefaultCommand {
         }
 
       } catch (IOException e) {
-        err(e.getMessage());
+        fatal(e.getMessage());
       }
     }
   }
@@ -364,7 +364,7 @@ public class InstallCommand extends DefaultCommand {
 
     // this is a best effort
     if (!exe.setExecutable(true, false)) {
-      err("Cannot set script 'node_modules/.bin/" + launcher + "'executable!");
+      fatal("Cannot set script 'node_modules/.bin/" + launcher + "'executable!");
     }
   }