+++ pm/src/main/java/io/reactiverse/es4x/commands/VscodeCommand.java
@@ -15,7 +15,6 @@
  */
 package io.reactiverse.es4x.commands;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import io.vertx.core.cli.CLIException;
 import io.vertx.core.cli.annotations.*;
 import io.vertx.core.spi.launcher.DefaultCommand;
@@ -27,10 +26,9 @@ import java.util.*;
 import static io.reactiverse.es4x.commands.Helper.*;
 
 @Name("vscode")
-@Summary("Play with your vscode project.")
+@Summary("Launcher for vscode project.")
 public class VscodeCommand extends DefaultCommand {
 
-	private static final ObjectMapper MAPPER = new ObjectMapper();
 	private boolean launcher;
 
 	@Option(longName = "launcher", shortName = "l", flag = true)
@@ -39,19 +37,47 @@ public class VscodeCommand extends DefaultCommand {
 		this.launcher = create;
 	}
 
-	private static void processLauncher(File json) throws IOException {
-		Map launch = MAPPER.readValue(json, Map.class);
-		if (!launch.containsKey("configurations")) {
+	private void processLauncher(File json) throws IOException {
+
+    final File pkg = new File(getCwd(), "package.json");
+
+    String app = "Launch";
+
+    if (pkg.exists()) {
+      Map pkgJson = read(pkg);
+      app = "Launch " + pkgJson.get("name");
+    }
+
+    Map launch = read(json);
+
+    if (!launch.containsKey("configurations")) {
 			launch.put("configurations", new ArrayList<>());
 		}
+
 		final List configurations = (List) launch.get("configurations");
-		Map<String, Object> config = new LinkedHashMap<String, Object>();
-		config.put("name", "Launch via npm");
+
+    // replace the launcher if already present
+    Object toRemove = null;
+
+    for (Object c : configurations) {
+      Map<String, Object> config = (Map<String, Object>) c;
+      if (app.equals(config.get("name"))) {
+        toRemove = c;
+        break;
+      }
+    }
+
+    if (toRemove != null) {
+      configurations.remove(toRemove);
+    }
+
+		Map<String, Object> config = new LinkedHashMap<>();
+		config.put("name", app);
 		config.put("type", "node");
 		config.put("request", "launch");
 		config.put("cwd", "${workspaceFolder}");
 		config.put("runtimeExecutable", "npm");
-		List<String> npmArgs = new ArrayList<String>();
+		List<String> npmArgs = new ArrayList<>();
 		npmArgs.add("run-script");
 		npmArgs.add("start");
 		npmArgs.add("--");
@@ -59,8 +85,15 @@ public class VscodeCommand extends DefaultCommand {
 		config.put("runtimeArgs", npmArgs);
 		config.put("port", "5858");
 		config.put("outputCapture", "std");
+		// server ready
+    Map<String, Object> serverReady = new LinkedHashMap<>();
+    serverReady.put("pattern", "Listening on port ([0-9]+)");
+    serverReady.put("uriFormat", "http://localhost:%s");
+    serverReady.put("action", "openExternally");
+    config.put("serverReadyAction", serverReady);
+
 		configurations.add(config);
-		MAPPER.writeValue(json, launch);
+		write(json, launch);
 	}
 
 	@Override
@@ -69,23 +102,23 @@ public class VscodeCommand extends DefaultCommand {
 
 		if (!vscodefile.exists()) {
 			new File(getCwd(), ".vscode").mkdirs();
-			try (InputStream in = InitCommand.class.getClassLoader()
+			try (InputStream in = VscodeCommand.class.getClassLoader()
 					.getResourceAsStream("META-INF/es4x-commands/vscode-launcher.json")) {
 				if (in == null) {
-					err("Cannot load vscode launcher.json template.");
+					fatal("Cannot load vscode launcher.json template.");
 				} else {
 					Files.copy(in, vscodefile.toPath());
 				}
 			} catch (IOException e) {
-				err(e.getMessage());
+				fatal(e.getMessage());
 			}
 		}
 
-		if (this.launcher) {
+		if (launcher) {
 			try {
 				processLauncher(vscodefile);
 			} catch (IOException e) {
-				err(e.getMessage());
+				fatal(e.getMessage());
 			}
 		}
 	}