+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] -
+- Removed `es4x-cli` as it has been replaced with `es4x-pm`
 
 ## [0.7.3] - 2019-03-22
 - Fixes bash script for debian like systems